feat: use decoded github provider private key
The GitHub Terraform provider requires the *content* of the private key to be
passed as an input in the provider configuration, rather than a pointer to a
file, despite the misleading `pem_file` parameter name. This PR updates the
workflow appropriately. The decoded private key is stored in `/tmp` and removed
at the end of the workflow to alleviate security concerns on long-lived runners.